(MODULES-6598) remove semantic_puppet dependency
Puppet 5.3.4 does some new i18n-fu and takes matters into its own hands when handling text_domains and so on, breaking the dependency on ye olde gettext-setup gem. semantic_puppet still uses the gettext-setup gem and when it tries to switch the text_domain to "master_domain", it does not exist. puppet_pot_generator depends on semantic_puppet, so the gem presumably pulls that dependency in and  things try to use it instead of the vendored semantic_puppet. Also, Puppet > 4.9.0 does not require semantic_puppet any longer, so we are killing it with fire. Here, anyway.